move SSEStreamResponse to vercel.py
